+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- [frontend]: Persist the filter selection (debug level and search criteria) in the Logs route.
+- [frontend]: Added version and two badges for bridgeMode and restartMode.
+- [frontend]: Frontend updated to 0.8.8.
 - [docker]: Added support for docker. The Matterbridge image is published on the docker hub.
 - [docker]: See the guidelines on https://github.com/Luligu/matterbridge?tab=readme-ov-file#Run-the-Docker-container-and-start-it.
 - [docker compose]: See the guidelines on https://github.com/Luligu/matterbridge?tab=readme-ov-file#Run-with-docker-compose.